--- macros/blueprint-macro/tests/invalid_cases/01_missing_params.rs
@@ -1,15 +1,0 @@
-use gadget_blueprint_macro::job;
-
-struct Bytes;
-impl Bytes {
-    fn new() -> Self {
-        Self
-    }
-}
-
-#[job(result(Bytes))]
-fn keygen(n: u16) -> Bytes {
-    Bytes::new()
-}
-
-fn main() {}
