+++ macros/blueprint-proc-macro/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "gadget-blueprint-macro"
+name = "gadget-blueprint-proc-macro"
 version = "0.1.0"
 authors.workspace = true
 edition.workspace = true
