+++ macros/blueprint-proc-macro/tests/invalid_cases/03_invalid_params_syntax.rs
@@ -1,0 +1,15 @@
+use gadget_blueprint_proc_macro::job;
+
+struct Bytes;
+impl Bytes {
+    fn new() -> Self {
+        Self
+    }
+}
+
+#[job(params(n t), result(Bytes))]
+fn keygen(n: u16, t: u8) -> Bytes {
+    Bytes::new()
+}
+
+fn main() {}
