+++ macros/blueprint-proc-macro/tests/invalid_cases/05_function_without_return.rs
@@ -1,4 +1,4 @@
-use gadget_blueprint_macro::job;
+use gadget_blueprint_proc_macro::job;
 
 struct Bytes;
 impl Bytes {
