+++ macros/blueprint-proc-macro/tests/invalid_cases/05_function_without_return.stderr
